Add [ESPXXX] commands be accessible from Serial / Telnet Thanks @bdring
Better Bluetooth integration
