+++ Grbl_Esp32/report.h
@@ -106,7 +106,7 @@
 #define CLIENT_COUNT    4 // total number of client types regardless if they are used
 
 // functions to send data to the user.
-void grbl_send(uint8_t client, char *text);
+void grbl_send(uint8_t client, const char *text);
 void grbl_sendf(uint8_t client, const char *format, ...);
 
 // Prints system status messages.
