+++ partitions/min_spiffs.csv
@@ -1,0 +1,7 @@
+# Name,   Type, SubType, Offset,  Size, Flags
+nvs,      data, nvs,     0x9000,  0x5000,
+otadata,  data, ota,     0xe000,  0x2000,
+app0,     app,  ota_0,   0x10000, 0x1E0000,
+app1,     app,  ota_1,   0x1F0000,0x1E0000,
+eeprom,   data, 0x99,    0x3D0000,0x1000,
+spiffs,   data, spiffs,  0x3D1000,0x2F000,
